cf workers tests use a createTestBillingContext
